renamed db package to config and changed NewPostFactory method to accept parameter
