+++ config/connection.go
@@ -1,4 +1,4 @@
-package db
+package config
 
 import (
 	"babalaas/stella-artois/model"
